* **New Features**
  * ESP32‑S3 target support (optional M5) with end‑to‑end build/flash tooling.
  * Enhanced MQTT client and field‑sync: TLS, QoS, health and backpressure metrics.
  * Audio subsystem: explicit read/write results, backpressure strategies, shutdown/diagnostics.
  * Configurable embedded Python runtime; new image types and offscreen renderer PNG capture + RGBA blit.

* **Improvements**
  * CI expanded to include ESP32 variants; clearer build/submodule messages.
  * README/CONTRIBUTING formatting, examples and docs updated.

* **Tests**
  * New unit and integration tests for MQTT, audio, Python runtime and related components.

+++ cpp/include/robotick/systems/auditory/CochlearTransform.h
@@ -1,4 +1,4 @@
-// Copyright Robotick
+// Copyright Robotick contributors
 // SPDX-License-Identifier: Apache-2.0
 //
 // CochlearTransform.h  (lean header: declarations only)
@@ -6,7 +6,7 @@
 #pragma once
 
 #include "robotick/api.h"
-#include "robotick/framework/common/FixedVector.h"
+#include "robotick/framework/containers/FixedVector.h"
 #include "robotick/systems/audio/AudioFrame.h"
 #include "robotick/systems/auditory/CochlearFrame.h"
 
@@ -144,7 +144,7 @@ namespace robotick
 		static int clamp_fft_bin_index(int bin_index);
 
 		// Denormal suppression for tiny floats.
-		static inline float zap_denorm(float value) { return (std::fabs(value) < 1e-30f) ? 0.0f : value; }
+		static inline float zap_denorm(float value) { return (fabsf(value) < 1e-30f) ? 0.0f : value; }
 	};
 
 } // namespace robotick
